Push of work on devphone.c
Devphone.c is getting very messy at this point, going to split it off into at least 3 different files soon. Malloc wrappers have been left in the code for now, since when I remove them the code segfaults when making a voice call. The cause of the problem doesn't seem to be in devphone.c; it might be in one of the other files we've modified, like port/main.c.